+++ packages/path-inputs/src/index.ts
@@ -5,3 +5,4 @@ export * from './hint';
 export * from './typography';
 export * from './tag';
 export * from './text-area-input';
+export * from './time-input';