+++ packages/path-inputs/src/typography.tsx
@@ -7,7 +7,11 @@ const classNames = require('classnames');
 export enum TypographyTypes {
   Paragraph = 'Paragraph',
   Label18 = 'Label18',
+  Label18Normal = 'Label18Normal',
+  Label15 = 'Label15',
   Label14 = 'Label14',
+  Label14White = 'Label14White',
+  Label14Gray = 'Label14Gray',
   Label13 = 'Label13',
   Label12 = 'Label12',
   Label12Black = 'Label12Black',
@@ -59,11 +63,29 @@ const useStyles = createUseStyles((theme: Theme) => ({
     lineHeight: 'normal',
     color: '#434a4b'
   },
+  Label18Normal: {
+    fontSize: '18px',
+    fontWeight: 'normal',
+    lineHeight: 'normal',
+    color: '#434a4b'
+  },
   label14: {
     fontSize: '14px',
     fontWeight: 'bold',
     lineHeight: 'normal',
     color: theme.colors.grayscale1
+  },
+  label14Gray: {
+    fontSize: '14px',
+    fontWeight: 'bold',
+    lineHeight: 'normal',
+    color: theme.colors.grayscale2
+  },
+  label14White : {
+    fontSize: '14px',
+    fontWeight: 'bold',
+    lineHeight: 'normal',
+    color: theme.colors.white
   },
   label13: {
     fontSize: '13px',
@@ -73,6 +95,12 @@ const useStyles = createUseStyles((theme: Theme) => ({
   },
   label12: {
     fontSize: '12px',
+    fontWeight: 'bold',
+    lineHeight: 'normal',
+    color: theme.colors.grayscale2
+  },
+  label15: {
+    fontSize: '15px',
     fontWeight: 'bold',
     lineHeight: 'normal',
     color: theme.colors.grayscale2
@@ -111,8 +139,12 @@ export const Typography = (props: TypographyProps) => {
         {
           [classes.fontStyle]: true,
           [classes.label18]: type === TypographyTypes.Label18,
+          [classes.Label18Normal]: type === TypographyTypes.Label18Normal,
           [classes.label14]: type === TypographyTypes.Label14,
+          [classes.label14Gray]: type === TypographyTypes.Label14Gray,
+          [classes.label14White]: type === TypographyTypes.Label14White,
           [classes.label13]: type === TypographyTypes.Label13,
+          [classes.label15]: type === TypographyTypes.Label15,
           [classes.label12]: type === TypographyTypes.Label12,
           [classes.label12Black]: type === TypographyTypes.Label12Black,
           [classes.label12White]: type === TypographyTypes.Label12White,