+++ packages/path-kaltura/src/kaltura-player.stories.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import {KalturaPlayer} from "./kaltura-player";
-import {KalturaPlayerManager} from "./kaltura-player-manager";
+import {KalturaPlayerProvider} from "./kaltura-player-provider";
 import {createUseStyles} from "@kaltura-react-ui-kits/path-theming";
 import { withKnobs } from '@storybook/addon-knobs';
 import {Button} from "@kaltura-react-ui-kits/path-inputs";
-import {useContext} from "react";
+import {useContext, useEffect, useState} from "react";
 import {KalturaPlayerContext} from "./kaltura-player-context";
 import '../kaltura-player.css';
 
@@ -33,7 +33,7 @@ export const Default: Story = () => {
   const classes = useStyle();
 
   return (
-    <KalturaPlayerManager config={{
+    <KalturaPlayerProvider config={{
       ks:ks,
       partnerId: partnerId,
       uiConfId: uiConfId,
@@ -44,7 +44,7 @@ export const Default: Story = () => {
                        autoplay={false}
                        onMediaLoaded={(entryId) => console.log(entryId)}/>
       </div>
-    </KalturaPlayerManager>
+    </KalturaPlayerProvider>
   )
 };
 
@@ -60,8 +60,8 @@ export const KalturaPlayerAutoPlay: Story = () => {
   const classes = useStyle();
 
   return (
-    <KalturaPlayerManager autoLoad={true}
-                          config={{
+    <KalturaPlayerProvider autoLoad={true}
+                           config={{
                             ks:ks,
                             partnerId: partnerId,
                             uiConfId: uiConfId,
@@ -71,7 +71,7 @@ export const KalturaPlayerAutoPlay: Story = () => {
         <KalturaPlayer entryId={entryId}
                        onMediaLoaded={(entryId) => console.log(entryId)}/>
       </div>
-    </KalturaPlayerManager>
+    </KalturaPlayerProvider>
   )
 };
 
@@ -87,8 +87,8 @@ export const KalturaPlayerWithoutAutoPlay: Story = () => {
   const classes = useStyle();
 
   return (
-    <KalturaPlayerManager autoLoad={true}
-                          config={{
+    <KalturaPlayerProvider autoLoad={true}
+                           config={{
                             ks:ks,
                             partnerId: partnerId,
                             uiConfId: uiConfId,
@@ -99,7 +99,7 @@ export const KalturaPlayerWithoutAutoPlay: Story = () => {
                        autoplay={false}
                        onMediaLoaded={(entryId) => console.log(entryId)}/>
       </div>
-    </KalturaPlayerManager>
+    </KalturaPlayerProvider>
   )
 };
 
@@ -117,8 +117,8 @@ export const kalturaPlayerErrorLoadingBundler: Story = () => {
   const classes = useStyle();
 
   return (
-    <KalturaPlayerManager autoLoad={true}
-                          config={{
+    <KalturaPlayerProvider autoLoad={true}
+                           config={{
                             ks:ks,
                             partnerId: partnerId,
                             uiConfId: uiConfId,
@@ -132,7 +132,7 @@ export const kalturaPlayerErrorLoadingBundler: Story = () => {
                        onMediaLoadingError={(error => console.log(error))}
         />
       </div>
-    </KalturaPlayerManager>
+    </KalturaPlayerProvider>
   )
 };
 
@@ -149,8 +149,8 @@ export const MultiplePlayersInPage: Story = () => {
   const classes = useStyle();
 
   return (
-    <KalturaPlayerManager autoLoad={true}
-                          config={{
+    <KalturaPlayerProvider autoLoad={true}
+                           config={{
                             ks:ks,
                             partnerId: partnerId,
                             uiConfId: uiConfId,
@@ -169,7 +169,7 @@ export const MultiplePlayersInPage: Story = () => {
                          onMediaLoaded={(entryId) => console.log(entryId)}/>
         </div>
       </>
-    </KalturaPlayerManager>
+    </KalturaPlayerProvider>
   )
 };
 
@@ -204,8 +204,8 @@ export const ManuallyLoadPlayerBundlerScripts: Story = () => {
 
   return (
     <>
-      <KalturaPlayerManager autoLoad={false}
-                            config={{
+      <KalturaPlayerProvider autoLoad={false}
+                             config={{
                               ks:ks,
                               partnerId: partnerId,
                               uiConfId: uiConfId,
@@ -219,7 +219,7 @@ export const ManuallyLoadPlayerBundlerScripts: Story = () => {
                            onMediaLoaded={(entryId) => console.log(entryId)}/>
           </div>
         </>
-      </KalturaPlayerManager>
+      </KalturaPlayerProvider>
     </>
   )
 };
@@ -233,19 +233,19 @@ ManuallyLoadPlayerBundlerScripts.story = {
 };
 
 
-export const TowIdenticalPlayerProviders: Story = () => {
+export const TwoIdenticalPlayerProviders: Story = () => {
   const classes = useStyle();
 
   return (
-    <KalturaPlayerManager autoLoad={true}
-                          config={{
+    <KalturaPlayerProvider autoLoad={true}
+                           config={{
                             ks:ks,
                             partnerId: partnerId,
                             uiConfId: uiConfId,
                             bundlerUrl: bunderlUrl
                           }}>
-      <KalturaPlayerManager autoLoad={true}
-                            config={{
+      <KalturaPlayerProvider autoLoad={true}
+                             config={{
                               ks:ks,
                               partnerId: partnerId,
                               uiConfId: uiConfId,
@@ -256,12 +256,12 @@ export const TowIdenticalPlayerProviders: Story = () => {
                          autoplay={false}
                          onMediaLoaded={(entryId) => console.log(entryId)}/>
         </div>
-      </KalturaPlayerManager>
-    </KalturaPlayerManager>
+      </KalturaPlayerProvider>
+    </KalturaPlayerProvider>
   )
 };
 
-TowIdenticalPlayerProviders.story = {
+TwoIdenticalPlayerProviders.story = {
   parameters: {
     docs: {
       storyDescription: `Having two identical Kaltura Player providers`
@@ -274,8 +274,8 @@ export const TowDifferentPlayerProviders: Story = () => {
   const classes = useStyle();
 
   return (
-    <KalturaPlayerManager autoLoad={true}
-                          config={{
+    <KalturaPlayerProvider autoLoad={true}
+                           config={{
                             ks:ks,
                             partnerId: partnerId,
                             uiConfId: uiConfId,
@@ -287,8 +287,8 @@ export const TowDifferentPlayerProviders: Story = () => {
                          autoplay={false}
                          onMediaLoaded={(entryId) => console.log(entryId)}/>
         </div>
-        <KalturaPlayerManager autoLoad={false}
-                              config={{
+        <KalturaPlayerProvider autoLoad={false}
+                               config={{
                                 ks:ks,
                                 partnerId: partnerId,
                                 uiConfId: uiConfId,
@@ -302,9 +302,9 @@ export const TowDifferentPlayerProviders: Story = () => {
                              onMediaLoaded={(entryId) => console.log(entryId)}/>
             </div>
           </>
-        </KalturaPlayerManager>
+        </KalturaPlayerProvider>
       </>
-    </KalturaPlayerManager>
+    </KalturaPlayerProvider>
   )
 };
 
@@ -335,6 +335,280 @@ KalturaPlayerWithNoProvider.story = {
   }
 };
 
+const PlayerSeekAndShow = (props: {playerId: string}) => {
+
+  const {playerId} = props;
+
+  const {getPlayerCurrentTime$, seek} = useContext(KalturaPlayerContext);
+  const [currentTime, setcurrentTime] = useState(0);
+
+  useEffect(() => {
+    if(!playerId) return;
+
+    getPlayerCurrentTime$(playerId).subscribe((currentTime) => {
+      setcurrentTime(currentTime);
+    })
+  }, [playerId]);
+
+  const seekTo = () => {
+    if(!playerId) return;
+    seek(playerId, {seekTo: Math.floor(currentTime + 10*1000), pause: false});
+  };
+
+  return (
+    <div>
+      <div>{`Player ${playerId} currentTime: ${currentTime}`}</div>
+      <Button label={'Add 10 seconds'} onClick={seekTo}/>
+    </div>
+  );
+};
+
+export const MultiplePlayersWithTimeAndSeekInPage: Story = () => {
+  const classes = useStyle();
+
+  const [playerAId, setPlayerAId] = useState('');
+  const [playerBId, setPlayerBId] = useState('');
+
+  const onPlayerALoad = (data: {entryId: string, playerId: string}) => {
+    setPlayerAId(data.playerId);
+  };
+
+  const onPlayerBLoad = (data: {entryId: string, playerId: string}) => {
+    setPlayerBId(data.playerId);
+  };
+
+
+  return (
+    <KalturaPlayerProvider autoLoad={true}
+                           config={{
+                             ks:ks,
+                             partnerId: partnerId,
+                             uiConfId: uiConfId,
+                             bundlerUrl: bunderlUrl
+                           }}>
+      <>
+        <div className={classes.playerContainer}>
+          <KalturaPlayer entryId={entryId}
+                         autoplay={false}
+                         onPlayerLoaded={onPlayerALoad}
+                         onMediaLoaded={(entryId) => console.log(entryId)}/>
+        </div>
+        <PlayerSeekAndShow playerId={playerAId}/>
+        <br/>
+        <div className={classes.playerContainer}>
+          <KalturaPlayer entryId={entryId}
+                         autoplay={false}
+                         onPlayerLoaded={onPlayerBLoad}
+                         onMediaLoaded={(entryId) => console.log(entryId)}/>
+        </div>
+        <PlayerSeekAndShow playerId={playerBId}/>
+      </>
+    </KalturaPlayerProvider>
+  )
+};
+
+MultiplePlayersWithTimeAndSeekInPage.story = {
+  parameters: {
+    docs: {
+      storyDescription: `Kaltura Player multiple instances`
+    }
+  }
+};
+
+const PlayerPlayAndPauseActions = (props: {playerId: string}) => {
+
+  const {playerId} = props;
+
+  const {getPlayerCurrentTime$, play, pause} = useContext(KalturaPlayerContext);
+  const [currentTime, setcurrentTime] = useState(0);
+
+  useEffect(() => {
+    if(!playerId) return;
+
+    getPlayerCurrentTime$(playerId).subscribe((currentTime) => {
+      setcurrentTime(currentTime);
+    })
+  }, [playerId]);
+
+  const playAction = () => {
+    if(!playerId) return;
+    play(playerId);
+  };
+
+  const pauseAction = () => {
+    if(!playerId) return;
+    pause(playerId);
+  };
+
+  return (
+    <div>
+      <div>{`Player ${playerId} currentTime: ${currentTime}`}</div>
+      <Button label={'play'} onClick={playAction}/>
+      <Button label={'pause'} onClick={pauseAction}/>
+    </div>
+  );
+};
+
+
+export const MultiplePlayersForPlayAndPauseActions: Story = () => {
+  const classes = useStyle();
+
+  const [playerAId, setPlayerAId] = useState('');
+  const [playerBId, setPlayerBId] = useState('');
+
+  const onPlayerALoad = (data: {entryId: string, playerId: string}) => {
+    setPlayerAId(data.playerId);
+  };
+
+  const onPlayerBLoad = (data: {entryId: string, playerId: string}) => {
+    setPlayerBId(data.playerId);
+  };
+
+
+  return (
+    <KalturaPlayerProvider autoLoad={true}
+                           config={{
+                             ks:ks,
+                             partnerId: partnerId,
+                             uiConfId: uiConfId,
+                             bundlerUrl: bunderlUrl
+                           }}>
+      <>
+        <div className={classes.playerContainer}>
+          <KalturaPlayer entryId={entryId}
+                         autoplay={false}
+                         onPlayerLoaded={onPlayerALoad}
+                         onMediaLoaded={(entryId) => console.log(entryId)}/>
+        </div>
+        <PlayerPlayAndPauseActions playerId={playerAId}/>
+        <br/>
+        <div className={classes.playerContainer}>
+          <KalturaPlayer entryId={entryId}
+                         autoplay={false}
+                         onPlayerLoaded={onPlayerBLoad}
+                         onMediaLoaded={(entryId) => console.log(entryId)}/>
+        </div>
+        <PlayerPlayAndPauseActions playerId={playerBId}/>
+      </>
+    </KalturaPlayerProvider>
+  )
+};
+
+MultiplePlayersForPlayAndPauseActions.story = {
+  parameters: {
+    docs: {
+      storyDescription: `Kaltura Player multiple instances for testing play and pause action streams`
+    }
+  }
+};
+
+
+const PlayerStateUpdateLabels = (props: {playerId: string}) => {
+
+  const {playerId} = props;
+
+  const {getPlayerState$} = useContext(KalturaPlayerContext);
+  const [currentState, setCurrentState] = useState('idle');
+
+  useEffect(() => {
+    if(!playerId) return;
+
+    getPlayerState$(playerId).subscribe((currentState) => {
+      setCurrentState(currentState);
+    })
+  }, [playerId]);
+
+  return (
+    <div>
+      <div>{`Player ${playerId} currentState: ${currentState}`}</div>
+    </div>
+  );
+};
+
+
+export const MultiplePlayersForPlayerStateUpdate: Story = () => {
+  const classes = useStyle();
+
+  const [playerAId, setPlayerAId] = useState('');
+  const [playerBId, setPlayerBId] = useState('');
+
+  const onPlayerALoad = (data: {entryId: string, playerId: string}) => {
+    setPlayerAId(data.playerId);
+  };
+
+  const onPlayerBLoad = (data: {entryId: string, playerId: string}) => {
+    setPlayerBId(data.playerId);
+  };
+
+
+  return (
+    <KalturaPlayerProvider autoLoad={true}
+                           config={{
+                             ks:ks,
+                             partnerId: partnerId,
+                             uiConfId: uiConfId,
+                             bundlerUrl: bunderlUrl
+                           }}>
+      <>
+        <div className={classes.playerContainer}>
+          <KalturaPlayer entryId={entryId}
+                         autoplay={false}
+                         onPlayerLoaded={onPlayerALoad}
+                         onMediaLoaded={(entryId) => console.log(entryId)}/>
+        </div>
+        <PlayerStateUpdateLabels playerId={playerAId}/>
+        <br/>
+        <div className={classes.playerContainer}>
+          <KalturaPlayer entryId={entryId}
+                         autoplay={false}
+                         onPlayerLoaded={onPlayerBLoad}
+                         onMediaLoaded={(entryId) => console.log(entryId)}/>
+        </div>
+        <PlayerStateUpdateLabels playerId={playerBId}/>
+      </>
+    </KalturaPlayerProvider>
+  )
+};
+
+MultiplePlayersForPlayerStateUpdate.story = {
+  parameters: {
+    docs: {
+      storyDescription: `Kaltura Player multiple instances for testing players state update`
+    }
+  }
+};
+
+
+const workshopEntry = '1_qk8sqm6v';
+const workshopKs = 'Yzk0NzE1MjZmZDMyZmI0NzFiZThjODQ1YjM3NmY0YjhiOGU4OGZmMHwxODI3NTUxOzE4Mjc1NTE7MTYxNzAyMTk4NTswOzE1ODcwMjE5ODUuOTkyMjtzaGFpLmFpbnZvbmVyQGthbHR1cmEuY29tO3N2aWV3OjFfcWs4c3FtNnYsdmlldzoxX3FrOHNxbTZ2LGxpc3Q6Kjs7';
+
+export const KalturaPlayerWorkshopForTesting: Story = () => {
+  const classes = useStyle();
+
+  return (
+    <KalturaPlayerProvider autoLoad={true}
+                           config={{
+                             ks:workshopKs,
+                             partnerId: partnerId,
+                             uiConfId: uiConfId,
+                             bundlerUrl: bunderlUrl
+                           }}>
+      <div className={classes.playerContainer}>
+        <KalturaPlayer entryId={workshopEntry}
+                       autoplay={false}
+                       onMediaLoaded={(entryId) => console.log(entryId)}/>
+      </div>
+    </KalturaPlayerProvider>
+  )
+};
+
+KalturaPlayerWorkshopForTesting.story = {
+  parameters: {
+    docs: {
+      storyDescription: `Kaltura Player workshop for testing`
+    }
+  }
+};
 
 export default {
   title: 'Kaltura/Player',