+++ packages/path-theming/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0](https://github.com/kaltura/path-design-system/compare/v2.3.0...v3.0.0) (2020-06-01)
+
+**Note:** Version bump only for package @kaltura-react-ui-kits/path-theming
+
+
+
+
+
+# [2.0.0](https://github.com/kaltura/path-design-system/compare/v1.4.1...v2.0.0) (2020-05-11)
+
+**Note:** Version bump only for package @kaltura-react-ui-kits/path-theming
+
+
+
+
+
+# [1.0.0](https://github.com/kaltura/path-design-system/compare/v0.7.0...v1.0.0) (2020-04-12)
+
+**Note:** Version bump only for package @kaltura-react-ui-kits/path-theming
+
+
+
+
+
+# [0.7.0](https://github.com/kaltura/path-design-system/compare/v0.6.3...v0.7.0) (2020-04-05)
+
+**Note:** Version bump only for package @kaltura-react-ui-kits/path-theming
+
+
+
+
+
 # [0.6.0](https://github.com/kaltura/path-design-system/compare/v0.5.19...v0.6.0) (2020-03-29)
 
 **Note:** Version bump only for package @kaltura-react-ui-kits/path-theming